requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-devtools==0.13.1
aiojobs==0.2.2
python-engineio==3.13.2
python-socketio==4.6.0
async-timeout==3.0.1
attrs==19.3.0
brotlipy==0.7.0
cchardet==2.1.6
cffi==1.14.0
chardet==3.0.4
devtools==0.5.1
idna==2.9
gino==1.0.1
gino-aiohttp==0.1.0
multidict==4.7.5
Pygments==2.6.1
pycares==3.1.1
pycparser==2.20
pylint==2.5.0
requests==2.23.0
yarl==1.4.2
watchgod==0.6
-e git+https://github.com/SheetWithoutShit/sws-core.git@v0.5.2#egg=core
